Fix module crashing if a Gate is patched into the Gate Input
Crash occurs on VCV Rack (just patch Fundamental VCO square output to the Gate input of the Y module, and then wait a couple minutes).
Crash also occurs on MetaModule

The `setAction` feature seems incomplete, as `clearAction` is never called anywhere in the code. I don't really understand what the action feature is supposed to do or what time the actions are supposed to be cleared.